Added a Dockerfile.debian that seems to build the right files, but was not tested
The motivation is a smaller Docker and a preparation for combining it within the building of a single Docker image.
Re: smaller Docker image: When doing it on the Fedora image (i.e. dnf --setopt=install_weak_deps=False  ), not all modules are available for dracut (in particular tpm2-tss).

This is a work in progress, that may remain untouched as it is not
really important and I don't have the time for that.

The builder file uses the workdir/fedora/ folder as the work directory,
simply because the dracut command in build.sh is supposed to be
absolutely identical, and as mentioned, it is a work in progress